python_ingest/requirements.txt
fastf1
pandas
mysql-connector-python[pure]
